+++ project/Dependencies.scala
@@ -18,7 +18,7 @@ import sbt._
 object Dependencies {
 
   object Versions {
-    val crossScala = Seq("2.11.7", "2.10.5")
+    val crossScala = Seq("2.11.7")
     val scalaVersion = crossScala.head
   }