+++ silhouette/app/com/mohiva/play/silhouette/impl/authenticators/JWTAuthenticator.scala
@@ -17,7 +17,6 @@ package com.mohiva.play.silhouette.impl.authenticators
 
 import com.atlassian.jwt.SigningAlgorithm
 import com.atlassian.jwt.core.writer.{ JsonSmartJwtJsonBuilder, NimbusJwtWriterFactory }
-import com.mohiva.play.silhouette._
 import com.mohiva.play.silhouette.api.Authenticator.Implicits._
 import com.mohiva.play.silhouette.api.exceptions._
 import com.mohiva.play.silhouette.api.repositories.AuthenticatorRepository
@@ -259,7 +258,7 @@ class JWTAuthenticatorService(
    * @return Some authenticator or None if no authenticator could be found in request.
    */
   override def retrieve(implicit request: RequestHeader): Future[Option[JWTAuthenticator]] = {
-    Future.from(Try(request.headers.get(settings.headerName))).flatMap {
+    Future.fromTry(Try(request.headers.get(settings.headerName))).flatMap {
       case Some(token) => unserialize(token)(settings) match {
         case Success(authenticator) => repository.fold(Future.successful(Option(authenticator)))(_.find(authenticator.id))
         case Failure(e) =>